pyproject.toml
[project]
name = "aigw_ct"
version = "1.4.0"
description = "Агент ЦТ"
authors = [
    { name = "Alexandr Stikhin", email = "aalstikhin@sberbank.ru" }
]
maintainers = [
    { name = "Alexandr Stikhin", email = "aalstikhin@sberbank.ru" }
]
requires-python = ">=3.12,<4.0"


[project.scripts]
aigw_ct_sh = "aigw_ct.__main__:main"

[tool.uv.sources]
nexus = { index = "nexus" }

[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "httpx>=0.27.2",
    "pylint>=3.3.1",
    "pytest>=8.4.0",
    "pytest_env>=1.1.5",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.11.13",
    "isort>=6.0.1",
    "black>=25.1.0"
]
prod = [
    "langgraph>=1.0.1",
    "mammoth>=1.11.0",
    "json-repair>=0.52.3",
    "rapidfuzz>=3.14.1",
    "sse-starlette>=3.0.2",
    "fastapi>=0.115.6",
    "asyncio>=3.4.3",
    "loguru>=0.7.3",
    "pydantic>=2.10.5",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "pytz>=2025.1",
    "langchain>=0.3.20",
    "langchain-community>=0.3.20",
    "langchain-gigachat>=0.3.9",
    "pdfminer>=20191125",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "markdown>=3.8.2",
    "pandas>=2.3.0",
    "docx>=0.2.4",
    "python-magic>=0.4.27",
    "PyPDF2>=3.0.1",
    "pdfminer-six>=20250506",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "aiohttp>=3.12.15",
    "websockets>=15.0.1",
]

[tool.isort]
profile = "black"
skip = [".venv", "venv", "env"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.black]
line-length = 120

[tool.ruff]
extend-select = ["A", "B", "C4", "C90", "I", "Q", "PERF", "PT", "RUF100", "UP", "W"]
line-length = 120
src = ["src"]
target-version = "py311"
exclude = [
    "src/aigw_ct/interfaces/gigachat"
]
ignore = [
    "Q000",
    "I001",
    "B008"
]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = """
    too-many-arguments,
    not-callable,
    missing-class-docstring,
    missing-module-docstring,
    too-few-public-methods,
    broad-except,
    import-outside-toplevel,
    missing-function-docstring
"""


[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
asyncio_mode = "auto"
addopts = "-s --verbose -W ignore::DeprecationWarning"